+++ src/query_builder.ts
@@ -1,9 +1,34 @@
-import {QueryRequest} from "./types";
 import {Query} from "@google-cloud/datastore";
+import {QueryRequest} from "./types";
 
 export function queryBuilder(query: Query, req?: QueryRequest): Query {
   if (req) {
-    // do something
+
+    if (req.namespace) {
+
+      query.namespace = req.namespace;
+    }
+
+    if (req.order) {
+      query.order(req.order.field, {
+        descending: req.order.descending,
+      });
+    }
+
+    if (req.filter && req.filter.forEach) {
+      req.filter.forEach(f =>
+          // @ts-ignore
+          query.filter(f.property, f.operator, f.value)
+      );
+    }
+
+    if (req.start) {
+      query.start(req.start);
+    }
+
+    if (req.limit && req.limit > 0) {
+      query.limit(req.limit);
+    }
   }
   return query;
 }