+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## v0.1.1
+
+* Update `Filter` with missing fields
+* Update internal `QueryBuilder` to map a `Query` from a `QueryRequest`
+* Expose `Datastore` on `BaseRepository` allowing a Repository to extends it 
+* Rename internal `Repostiory` interface to `CrudRepository`
 
 ## v0.1.0 - Initial release